% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{MCMC}
\alias{MCMC}
\title{MCMC sampling for count regression with a constant mean, latent autocorrelated errors, and missing data}
\usage{
MCMC(
  response,
  datadist = dpois,
  antilink = exp,
  beta0ini,
  sigmaini = NULL,
  sigCARini,
  rhoini,
  zini,
  rhogrid,
  detsig,
  SigiList,
  eff_mat_offset = NULL,
  ind_samp,
  ind_miss,
  sample_sigma = FALSE,
  sample_sigCAR = TRUE,
  sample_rho = TRUE,
  use_trunc = FALSE,
  trunc_bound = NULL,
  nMCMC = 1000,
  n_z_iters = 10,
  n_thin = 1,
  samp_sd = 0.01,
  miss_sd = 0.01,
  beta0_sd = 0.1,
  sigma_sd = 0.1,
  sigCAR_sd = 0.02,
  rho_indxpm = 3
)
}
\arguments{
\item{response}{vector with response variable.}

\item{datadist}{probability density function for data.  It must be parameterized with a mean parameter, and a variance parameter.  If there is no variance parameter, the argument must still be there, and just evaluated to NULL.}

\item{antilink}{function to change mean (linear model) on link scale to mean for datadist distribution.}

\item{beta0ini}{initialize beta0}

\item{sigmaini}{initialize sigma for datadist, if it has one.}

\item{sigCARini}{initialize sigCAR}

\item{rhoini}{initialize rho. It must be one of the values in rhogrid}

\item{zini}{initialize latent autocorrelated errors z}

\item{rhogrid}{vector of rho values for lookup table}

\item{detsig}{vector of determinant values corresponding to rhogrid}

\item{SigiList}{list of inverse covariances, in sparse matrix form,}

\item{eff_mat_offset}{matrix of samples from posterior distribution for effort.  Default is NULL, in which case no offset is used}

\item{ind_samp}{vector of indexes of sampled grid cells}

\item{ind_miss}{vector of indexes of unsampled grid cells}

\item{sample_sigma}{Logical.  Should sigma parameter be sampled (TRUE) or held at initial values (FALSE).}

\item{sample_sigCAR}{Logical.  Should sigCAR parameter be sampled (TRUE) or held at initial values (FALSE).}

\item{sample_rho}{Logical.  Should rho parameter be sampled (TRUE) or held at initial values (FALSE).}

\item{use_trunc}{Logical.  Should truncation for z (latent autocorrelated errors) be used.  If true, truncated to plus and minus the log of the largest observed value.}

\item{nMCMC}{number of MCMC samples to retain}

\item{n_z_iters}{number of latent autocorrelated errors (z) proposals per further MCMC updating}

\item{n_thin}{number of MCMC samples per retained MCMC sample}

\item{samp_sd}{tuning parameter for width of independent uniform proposal for Metropolis step of sampled z-values}

\item{miss_sd}{tuning parameter for width of independent uniform proposal for Metropolis step of missing z-values}

\item{beta0_sd}{tuning parameter for variance of independent normal proposal for Metropolis step of beta0}

\item{sigma_sd}{tuning parameter for variance of independent normal proposal for Metropolis step of sigma}

\item{sigCAR_sd}{tuning parameter for variance of independent normal proposal for Metropolis step of sigCAR}

\item{rho_indxpm}{tuning parameter for number of adjacent grid values, plus and minus, of proposal for Metropolis/Hastings step of rho}
}
\value{
a list of MCMC samples of the posteriors and acceptance rates of Metropolis proposals
}
\description{
MCMC sampling for count regression with a constant mean, latent autocorrelated errors, and missing data
}
\author{
Jay Ver Hoef
}
