% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfun_probability.R
\name{probability}
\alias{is.probability}
\alias{p}
\alias{probability}
\title{probability transfun}
\usage{
probability(value)

p(value)

is.probability(x)
}
\arguments{
\item{value}{a numeric between 0 and 1 representing a probability}

\item{x}{an object to be tested as a probability transfun object}
}
\description{
Create a transfun object representing a probability of
 transition between states. Typically used inside a call to
 \code{\link{transition}}
}
\details{
\code{p()} is a shorthand for \code{probability()}.
}
\examples{
# these are equivalent
prob <- probability(0.2)
prob <- p(0.2)

is.probability(prob)
}

