% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfun_rate.R
\name{rate}
\alias{is.rate}
\alias{r}
\alias{rate}
\title{rate transfun}
\usage{
rate(value)

r(value)

is.rate(x)
}
\arguments{
\item{value}{a numeric greater than 0 representing a rate}

\item{x}{an object to be tested as a rate transfun object}
}
\description{
Create a transfun object representing a rate of transition
  between states - e.g. an expected number of offspring generated into one
  state from another. Typically used inside a call to
  \code{\link{transition}}
}
\details{
\code{r()} is a shorthand for \code{rate()}.
}
\examples{
# these are equivalent
rate <- rate(0.2)
rate <- r(0.2)

is.rate(rate)
}

