% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfun-class.R
\name{as.transfun}
\alias{as.transfun}
\title{create a transition function}
\usage{
as.transfun(fun, type = c("probability", "rate"))
}
\arguments{
\item{fun}{an R function describing the transition. This must take only one
argument: \dots, and return a single numeric value. The function may make
use of a limited set of objects which will be passed via \dots. These are
described in \code{details}}

\item{type}{what type of transition this function represents, a probabiltiy
or a rate}
}
\description{
A utility function to enable users to create bespoke transition
  functions (\code{transfun} objects) for use in \code{transition}s.
}
\details{
\code{fun} must take only the argument \dots, but other objects may
  be passed to the function via this argument. These objects are:
  \code{population}, a named numeric vector giving the number of individuals
  of each stage *within the patch*; \code{area} a single numeric value giving
  the area of the patch in square kilometres; \code{features} a named numeric
  vector containing miscellaneous features of the habitat patch, such and
  measures of patch quality or environmental variables. See examples for an
  illustration of how to these objects.
}
\examples{
# a very simple (and unnecessary, see ?p) transfun
fun <- function(...) 0.3
prob0_3 <- as.transfun(fun, type = 'probability')

# a density-dependent probability (population and area are passed via the
# dots)
dd_fun <- function (...) {
    adult_density <- population['adult'] / area
    p <- sqrt(1 / adult_density)
    return (p)
}
dd_prob <- as.transfun(dd_fun, type = 'probability')


}

