% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfun-class.R
\name{transfun}
\alias{is.transfun}
\alias{print.transfun}
\alias{transfun}
\title{transfun objects}
\usage{
is.transfun(x)

\method{print}{transfun}(x, ...)
}
\arguments{
\item{x}{a transfun object to print or an object to test as a transfun object}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
utility functions for the \code{transfun} class. \code{transfun}
  objects are created by functions such as \code{\link{probability}}.
}
\examples{
prob <- p(0.3)
is.transfun(prob)

prob
}

