% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfun-class.R
\name{transfun}
\alias{is.transfun}
\alias{transfun}
\title{transfun objects}
\usage{
is.transfun(x)
}
\arguments{
\item{a}{two-sided formula identifying the states between which the
transition occurs}

\item{transfun}{a \code{\link{transfun}} object quantifying the transition.}
}
\description{
creates a \code{transition} object, encoding a transition
  between two states. E.g. the probability of a seed germinating, or of an
  individual surviving in each time step
}
\examples{
prob <- p(0.3)
is.transfun(prob)
}

