% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition-class.R
\name{transition}
\alias{is.transition}
\alias{print.transition}
\alias{tr}
\alias{transition}
\title{transition objects}
\usage{
transition(formula, transfun)

tr(formula, transfun)

is.transition(x)

\method{print}{transition}(x, ...)
}
\arguments{
\item{formula}{a two-sided formula identifying the states between which the
transition occurs}

\item{transfun}{a \code{\link{transfun}} object quantifying the transition.}

\item{x}{an object to print or test as a transition object}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
creates a \code{transition} object, encoding a transition
  between two states. E.g. the probability of a seed germinating, or of an
  individual surviving in each time step
}
\details{
\code{tr} is just a shorthand for \code{transition}
}
\examples{
# 50/50 chance of a larva emerging from an egg
hatching <- tr(lava ~ egg, p(0.5))

# three eggs laid per adult per time step
fecundity <- tr(egg ~ adult, r(3))

# 0.1 probability of a larva pupating into an adult
pupa <- tr(adult ~ larva, p(0.1))

# print method
print(pupa)
}

